完成加入購物車及最愛按鈕
5/17

+++ src/main/java/tw/com/eeit162/meepleMasters/jim/mall/service/ProductService.java
@@ -2,18 +2,27 @@ package tw.com.eeit162.meepleMasters.jim.mall.service;
 
 import java.io.BufferedInputStream;
 import java.io.IOException;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import javax.transaction.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import tw.com.eeit162.meepleMasters.jack.model.bean.Member;
+import tw.com.eeit162.meepleMasters.jim.mall.model.bean.FavoriteGame;
 import tw.com.eeit162.meepleMasters.jim.mall.model.bean.Product;
+import tw.com.eeit162.meepleMasters.jim.mall.model.bean.ShoppingCart;
+import tw.com.eeit162.meepleMasters.jim.mall.model.dao.FavoriteGameDAO;
 import tw.com.eeit162.meepleMasters.jim.mall.model.dao.ProductDAO;
+import tw.com.eeit162.meepleMasters.jim.mall.model.dao.ShoppingCartDAO;
+import tw.com.eeit162.meepleMasters.jim.mall.model.dto.ProductDTO;
 
 @Service
 public class ProductService {
@@ -21,13 +30,55 @@ public class ProductService {
 	@Autowired
 	private ProductDAO pDAO;
 
-	public Page<Product> findAllProduct(Integer page, Integer count) {
+	@Autowired
+	private ShoppingCartDAO scDAO;
 
+	@Autowired
+	private FavoriteGameDAO fgDAO;
+
+	public void multiConditionQuery(Integer page, Integer count, Integer mID) {
+		PageRequest pageRequest = PageRequest.of(page - 1, count);
+		
+		Page<Product> productPerPage = pDAO.findAll(pageRequest);
+		
+		List<Product> content = productPerPage.getContent();
+		
+		content.stream().filter(p->p.getProductPlayTime().contains(""));
+		
+		
+	}
+
+	public PageImpl<ProductDTO> findAllProduct(Integer page, Integer count, Integer mID) {
 		PageRequest pageRequest = PageRequest.of(page - 1, count);
 
-		Page<Product> productPage = pDAO.findAll(pageRequest);
+		Page<Product> productPerPage = pDAO.findAll(pageRequest);
 
-		return productPage;
+		List<Product> content = productPerPage.getContent();
+		List<ProductDTO> collect = content.stream().map(p -> {
+			ProductDTO pDTO = new ProductDTO(p);
+			if (mID != null) {
+				List<ShoppingCart> currentMemberShoppingCarts = scDAO.findByMember(new Member(mID));
+				for (ShoppingCart sc : currentMemberShoppingCarts) {
+					if (sc.getProduct().equals(p)) {
+						pDTO.setIsInCart(true);
+						break;
+					}
+				}
+				List<FavoriteGame> currentMemberFavoriteGames = fgDAO.findByMember(new Member(mID));
+				for (FavoriteGame fg : currentMemberFavoriteGames) {
+					if (fg.getProduct().equals(p)) {
+						pDTO.setIsFavorite(true);
+						break;
+					}
+				}
+			}
+			return pDTO;
+		}).collect(Collectors.toList());
+
+		PageImpl<ProductDTO> pageImpl = new PageImpl<>(collect, productPerPage.getPageable(),
+				productPerPage.getTotalElements());
+
+		return pageImpl;
 	}
 
 	public Product findProductById(Integer id) {
@@ -52,8 +103,13 @@ public class ProductService {
 		return pDAO.save(p);
 	}
 
-	public void deleteProductById(Integer id) {
-		pDAO.deleteById(id);
+	public String deleteProductById(Integer id) {
+		Product product = pDAO.findById(id).get();
+		if (product != null) {
+			pDAO.deleteById(id);
+			return "刪除成功";
+		}
+		return "刪除失敗";
 	}
 
 	@Transactional
